removed unused methods from room repo

+++ ehotelservices/src/main/java/org/com/ehotel/service/room/RoomServiceImp.java
@@ -61,6 +61,9 @@ public class RoomServiceImp implements RoomService{
 
     @Override
     public void deleteRoom(Integer id) {
+        if(!roomRepository.existsById(id)) {
+            throw new AppEntityNotFoundException("Room not found with id: " + id);
+        }
         log.info("Deleting room with id: " + id);
         roomRepository.deleteById(id);
     }
